% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mft_test.R, R/ttest_berk.R
\name{compute_statistic}
\alias{compute_statistic}
\title{Takes two fitted sme models and returns
a moderated functional t-test statistic
after Berk.
funktional.norm
x, y curves
p = 2 is euclidean norm
by is stepsize of approximation}
\usage{
compute_statistic(fit1, fit2, points)

compute_statistic(fit1, fit2, points)
}
\description{
Takes two fitted sme models and returns
a moderated functional t-test statistic
after Berk.
funktional.norm
x, y curves
p = 2 is euclidean norm
by is stepsize of approximation

Takes two fitted sme models and returns
a moderated functional t-test statistic
after Berk.
funktional.norm
x, y curves
p = 2 is euclidean norm
by is stepsize of approximation
}

