% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv.lambda2.R
\name{cv.lambda2}
\alias{cv.lambda2}
\title{TODO}
\usage{
cv.lambda2(obslist, timelist, N = NULL, nbas = NULL, check.na = TRUE,
  trace = F, seed, ncores = 1)
}
\arguments{
\item{obslist}{TODO}

\item{timelist}{TODO}

\item{N}{TODO}

\item{nbas}{TODO}

\item{check.na}{TODO}

\item{trace}{TODO}

\item{seed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO

TODO
}
\note{
TODO
}
\examples{
TODO
}
\author{
TODO
}
\references{
TODO
}
\seealso{
TODO
}
\keyword{nonparametric}

