% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_test.R
\name{gam_test}
\alias{gam_test}
\title{Test based on additive, varying coefficients model}
\usage{
gam_test(data, group, timepoints, id, plot = FALSE)
}
\arguments{
\item{data}{[numeric vector] with the values of the response evaluated at the times in 'timepoints' argument.}

\item{group}{[numeric vector] of 1s and 2s as the grouping variable.}

\item{timepoints}{[numeric vector] with time points corresponding to measurements in 'data'.}

\item{id}{[numeric vector] with integers corresponding to the individual curves.}

\item{points}{[Integer] number of points used when discretizing functions.}

\item{nboot}{[Integer] number of Boostrap repetitions.}
}
\description{
Test based on additive, varying coefficients model
}

