% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_test.R
\name{ji_test}
\alias{ji_test}
\title{Joint confidence intervals test}
\usage{
ji_test(data, group, timepoints, id, method = c("spline", "rawdata"),
  points = 201, nboot = 1000, ncpus = 1, lambda = NULL, plot = FALSE)
}
\arguments{
\item{data}{[numeric vector] with the values of the response evaluated at the times in 'timepoints' argument.}

\item{group}{[numeric vector] of 1s and 2s as the grouping variable.}

\item{timepoints}{[numeric vector] with time points corresponding to measurements in 'data'.}

\item{id}{[numeric vector] with integers corresponding to the individual curves.}

\item{points}{[Integer] number of points used when discretizing functions.}

\item{nboot}{[Integer] number of Boostrap repetitions.}

\item{ncpus}{[Integer] number of CPUs to use. For greater than 1, the Bootstrapping is parallelized.}

\item{plot}{[Logical] should a plot of the confidence bands be shown?}
}
\description{
Joint confidence intervals test
}

