% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Internal function
takes a function and other parameters
returns simulated data, with certain shift, variance, and if
paired = TRUE also paired data with certain correlation}
\usage{
make_data(shift, n, sd1, sd2, ntimepoints, n_NA, type = c("vertshift",
  "horshift", "slopeshift", "allshift"), f = NULL, f2 = NULL)
}
\description{
Internal function
takes a function and other parameters
returns simulated data, with certain shift, variance, and if
paired = TRUE also paired data with certain correlation
}

