% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mft_test.R
\name{mft_test}
\alias{mft_test}
\title{Functional t-test for mean curves from two spline mixed effects ('sme') models}
\usage{
mft_test(data, group, timepoints, id, points = 500, nboot = 1000,
  ncpus = 1, ...)
}
\arguments{
\item{data}{[vector] with the response variable, i.e. values of some blood marker}

\item{group}{[vector] with grouping information. Has to be same length as data and
contain two groups.}

\item{timepoints}{[vector] with timepoints of measurments of the values in data}

\item{id}{[vector] with index for individuals}

\item{points}{[Integer] number of grid points used for discretizing the functions when integrating.}

\item{nboot}{[Integer] number of Bootstrap samples to use for estimating distribution of test statistic under H0.}

\item{ncpus}{[Integer] number of CPUs to use, parallelization for > 2.}
}
\value{
pvalue for group differences
}
\description{
Testing hypothesis mean(grp1) = mean(grp2).
}

