% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_p_min.R
\name{p_min}
\alias{p_min}
\title{Function to compare groups using the minimum of point-wise t-tests at the measurement times of 
functions.
Used internally in simulation study of TPDT, usage not recommended.}
\usage{
p_min(data, group, time, id)
}
\arguments{
\item{data}{[numeric vector] with the values of the response evaluated at the times in 'timepoints' argument.}

\item{group}{[numeric vector] of 1s and 2s as the grouping variable.}

\item{id}{[numeric vector] with integers corresponding to the individual curves.}

\item{timepoints}{[numeric vector] with time points corresponding to measurements in 'data'.}
}
\description{
Function to compare groups using the minimum of point-wise t-tests at the measurement times of 
functions.
Used internally in simulation study of TPDT, usage not recommended.
}

