% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.TPDT.R
\name{plot.TPDT}
\alias{plot.TPDT}
\title{Plot the outcome of a TPDT object}
\usage{
\method{plot}{TPDT}(x, y, plottype, ...)
}
\arguments{
\item{x}{object of class \code{TPDT}}

\item{y}{not used}

\item{plottype}{type of plot to be produced}

\item{...}{further arguments to be passed to the chosen plot function}
}
\description{
Plot an object of class \code{TPDT}. The function allows the user to select a plot type and thus providing the flexibility to choose which aspect of data to be graphically presented.
}
\details{
\code{plottype==1} (default) produces a histogram of the resampled test statistics and
displays the corresponding p-value of the test in the upper right of the plot window.

\code{plottype==2} presents the resampled test statistics as a horizontal boxplot and
again the p-value is displayed.

\code{plottype==3} presents the data, single steps of the computation workflow (difference
curves, mean of difference curves and standard deviation of difference curves) and  histogram
of resampled test statistics as well as the p-value and information about the rejection of the null
hypothesis.
\code{plottype==3} might be difficult for a large number of observations.
}
\examples{
f <- function(x) 2 * x * sin(x) + 10
# simulate paired data from two groups with underlying function f
simdata <- make_data(f = f, shift = 5, n = 5, sd1 = .5, sd2 = .5,
ntimepoints = 10, type = "shift")

# run test
res <- TPDT(simdata, B = 100)

# call plot on result of the test
plot(res, plottype = 1)
plot(res, plottype = 2)
plot(res, plottype = 3)
}
\author{
Ivan Kondofersky
}
\references{
To be added after acceptance of publication.
}
\seealso{
\code{\link{TPDT}}, \code{\link{summary.TPDT}}, \code{\link{print.TPDT}}
}

