% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data}
\usage{
prep_data(df, nbas, lambda, deriv)
}
\arguments{
\item{df}{[data.frame] with columns "data", "time", "group" and "id".}

\item{nbas}{[integer] number of basis functions to use.}

\item{deriv}{[integer] number of derivative to analyse.}
}
\value{
[list] With components "func1" and "func2" corresponding to the functional data for the
to groups to be compared. Both objects are of class "fd".
}
\description{
Internal function to make functional data out of the raw input data.frame.
}

