% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.basis.na.R
\name{smooth.basis.na}
\alias{smooth.basis.na}
\title{Extension of smooth.basis for some NA observations.}
\usage{
smooth.basis.na(argvals, y, fdParobj, ...)
}
\description{
Only for internal use in TPDT.
See smooth.basis for argument descriptions.
}

