% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPDT-package.R, R/TPDT.R
\docType{package}
\encoding{UTF-8}
\name{TPDT}
\alias{TPDT}
\alias{TPDT-package}
\alias{TPDT.data.frame}
\alias{TPDT.numeric}
\title{TPDT}
\usage{
TPDT(x, ...)

\method{TPDT}{numeric}(data, group, time, id, deriv = 0, B = 100,
  lambda = NULL, ncores = 1, ...)

\method{TPDT}{data.frame}(data, deriv = 0, B = 100, lambda = NULL,
  ncores = 1, nbas = NULL, ...)
}
\arguments{
\item{data}{either a [data.frame] with columns named "data", "group", "time" and "id" analogously to 
the use of \code{TPDT.numeric}. If using \code{TPDT.numeric}, a vector with same order as group, time and id vector.}

\item{group}{[numeric/factor] vector with grouping variable. A character vector will be 
converted in to numeric 1/2 internally.}

\item{time}{[numeric] vector of time points, has to be the same order as data, group and id.}

\item{id}{[numeric] vector of pairings. Each unique value of id should be present in both groups.}

\item{deriv}{[integer] If 1, test is conducted on first derivative of functional data. If 2, tests second derivative and so on.}

\item{lambda}{[numeric] Smoothing parameter to use, if known in advance. If NULL, it is computed by cross validation.}

\item{ncores}{[integer] Number of CPUs to use for parallization.}
}
\description{
The package provides the necessary tools to compute a TPDT for a general dataset.
The data required are paired, time-resloved observations. The output of the statistical test
allows to decide whether there are significant differences between both groups and to quantify
the impact of those differences. The package is extended by generic functions which may be
applied on the computed test object and help visualize the test results. Additionally a
function which allows to sample random curves around a given smooth curve and a cross
validation function are presented to the user.

Main function to perform a TPDT for a given dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Method for column wise input

\item \code{data.frame}: Method for data.frame input
}}
\examples{
# simulated random data
df <- data.frame(data = rnorm(60), expand.grid(time=1:10, group=1:2, id=1:3)) 
# apply test
res <- TPDT(df)
# print result
res
# plotting method
plot(res)
# illustration of concept
plot(res, plottype = 3)
}

