% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttest_berk.R
\name{ttest_berk}
\alias{ttest_berk}
\title{Functional t-test for mean curves from two spline mixed effects ('sme') models}
\usage{
ttest_berk(data, group, timepoints, id, points = 500, prob = 0, NAs = 0,
  nboot = 1000, method = c("non-parametric", "parametric"), ncpus = 1,
  ...)
}
\arguments{
\item{data}{[vector] with the response variable, i.e. values of some blood marker}

\item{group}{[vector] with grouping information. Has to be same length as data and
contain two groups.}

\item{timepoints}{[vector] with timepoints of measurments of the values in data}

\item{id}{[vector] with index for individuals}
}
\value{
pvalue for group differences
}
\description{
Functional t-test for mean curves from two spline mixed effects ('sme') models
}

