% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttest_crain.R
\name{ttest_crain}
\alias{ttest_crain}
\title{Test for group difference in functional data.}
\usage{
ttest_crain(data, group, timepoints, id, method = c("spline", "coef",
  "rawdata"), points = 201, nboot = 1000, ncpus = 1, lambda = NULL)
}
\arguments{
\item{data}{[vector] with the response variable, i.e. values of some blood marker}

\item{group}{[vector] with grouping information. Has to be same length as data and
contain two groups.}

\item{timepoints}{[vector] with timepoints of measurments of the values in data}

\item{id}{[vector] with index for individuals}

\item{method}{[choice] Currently only "non-parametric"-Version of the test available.}

\item{points}{[numeric] resolution of grid when discretisizing.}

\item{nboot}{[numeric] Number of bootstrap iterations.}

\item{ncpus}{[Integer] number of cores to use in parallel for bootstrap.}

\item{lambda}{[numeric] provide smoothing parameter. If NULL, an optimal choice is computed
through cross-validation.}
}
\value{
List with components: \itemize{
\item p: pvalue for group differences

\item fit1: fda object for group 1

\item fit2: fda object for group 2

\item lambda: smoothing parameter used for fitting.

}
}
\description{
Uses joint confidance bands for the difference curves of two groups.
Adapted from Crainiceanu, C. (2012).
}
\examples{
f <- function(x) 2 * x * sin(x) + 10
# simulate paired data from two groups with underlying function f
simdata <- make_data(f = f, shift = 5, n = 5, sd1 = .5, sd2 = .5,
                     ntimepoints = 10, type = "shift")
# run test
res <-  ttest_crain(data = simdata$data, group = simdata$group, timepoints = simdata$time, 
                   id = simdata$id, method = "spline", nboot = 1000)
}

