% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttest_berk.R
\name{compute_statistic}
\alias{compute_statistic}
\title{Takes two fitted sme models and returns
a moderated functional t-test statistic
after Berk.}
\usage{
compute_statistic(fit1, fit2, points)
}
\description{
Takes two fitted sme models and returns
a moderated functional t-test statistic
after Berk.
}

